pdf-extractor/requirements.txt
pypdf==4.2.0
pydantic>=2.7
openai>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.1
reportlab>=4.2.2
